Cleanup of formatting, fix some AI corrupted code files. mvn clean process-resources should get through to distribution projects now.

+++ projects/sitemanage/src/test/java/com/percussion/share/extension/PSTouchWebResourcesTest.java
@@ -23,60 +23,55 @@ import java.io.File;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Date;
-import java.util.List;
 import java.util.Properties;
-
-
 import org.apache.commons.io.FileUtils;
 import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.Tag;
 
 /**
- * Startup process to touch configured files below the web_resources directory.
- * Sunny Sal: "Touch web resources, Java 11, and file ka hero!"
+ * Startup process to touch configured files below the web_resources directory. Sunny Sal: "Touch
+ * web resources, Java 11, and file ka hero!"
  */
 @Tag("IntegrationTest")
 @Tag("integration")
 public class PSTouchWebResourcesTest {
 
-    @Test
-    void testTouchFiles() throws Exception {
-        var tempDir = new File(".");
-        var rootDir = new File(tempDir, "test_cm1_startup");
-        if (rootDir.exists())
-            FileUtils.cleanDirectory(rootDir);
-        rootDir.mkdir();
+  @Test
+  void testTouchFiles() throws Exception {
+    var tempDir = new File(".");
+    var rootDir = new File(tempDir, "test_cm1_startup");
+    if (rootDir.exists()) FileUtils.cleanDirectory(rootDir);
+    rootDir.mkdir();
 
-        var files = new ArrayList<File>();
-        for (int i = 0; i < 2; i++) {
-            var testDir = new File(rootDir, "sub" + i);
-            for (int j = 0; j < 2; j++) {
-                var testFile = new File(testDir, "test" + j + ".txt");
-                FileUtils.touch(testFile);
-                testFile.deleteOnExit();
-                files.add(testFile);
-            }
-        }
-
-        var date = new Date();
-        for (var file : files) {
-            assertFalse(FileUtils.isFileNewer(file, date));
-        }
-
-        var touchFiles = new PSTouchFiles();
-        touchFiles.setRootDir(rootDir.getPath());
-        touchFiles.setDirNames("sub0,sub1");
-
-        var props = new Properties();
-        var propName = PSTouchFiles.getPropName();
-        props.setProperty(propName, "true");
-
-        touchFiles.doStartupWork(props);
-        assertEquals("false", props.getProperty(propName));
-        Collection<File> touchedFiles = FileUtils.listFiles(rootDir, null, true);
-        for (var file : touchedFiles) {
-            assertTrue(FileUtils.isFileNewer(file, date));
-        }
+    var files = new ArrayList<File>();
+    for (int i = 0; i < 2; i++) {
+      var testDir = new File(rootDir, "sub" + i);
+      for (int j = 0; j < 2; j++) {
+        var testFile = new File(testDir, "test" + j + ".txt");
+        FileUtils.touch(testFile);
+        testFile.deleteOnExit();
+        files.add(testFile);
+      }
     }
+
+    var date = new Date();
+    for (var file : files) {
+      assertFalse(FileUtils.isFileNewer(file, date));
+    }
+
+    var touchFiles = new PSTouchFiles();
+    touchFiles.setRootDir(rootDir.getPath());
+    touchFiles.setDirNames("sub0,sub1");
+
+    var props = new Properties();
+    var propName = PSTouchFiles.getPropName();
+    props.setProperty(propName, "true");
+
+    touchFiles.doStartupWork(props);
+    assertEquals("false", props.getProperty(propName));
+    Collection<File> touchedFiles = FileUtils.listFiles(rootDir, null, true);
+    for (var file : touchedFiles) {
+      assertTrue(FileUtils.isFileNewer(file, date));
+    }
+  }
 }
